+++ scripts/commands/encrypt.sh
@@ -12,7 +12,7 @@ It uses your gpg credentials to encrypt .yaml file. If the file is already
 encrypted, look for a decrypted file and encrypt that to .yaml.
 This allows you to first decrypt the file, edit it, then encrypt it again.
 
-You can use plain sops to encrypt - https://github.com/mozilla/sops
+You can use plain sops to encrypt - https://github.com/getsops/sops
 
 Example:
   $ helm secrets enc <SECRET_FILE_PATH>