+++ CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- `secrets+*-import://` added syntax checks 
+- `secrets+*-import://` added syntax checks
+
+### Fixed
+- Multiline parameter values 
 
 ## [4.5.0] - 2023-08-14