+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `secrets+*-import://` added syntax checks
 
 ### Fixed
-- Multiline parameter values 
+- Multiline parameter values
+- Change all mozilla/sops to getsops/mozilla (https://github.com/getsops/sops/issues/1246)
 
 ## [4.5.0] - 2023-08-14
 
@@ -164,7 +165,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.10.0] - 2021-11-05
 ### Added
-- Add [age](https://github.com/mozilla/sops#encrypting-using-age) support for downloader plugin syntax.
+- Add [age](https://github.com/getsops/sops#encrypting-using-age) support for downloader plugin syntax.
 
 ### Changed
 - Improvements to the ArgoCD integration documentation.